+++ src/types/calendar.ts
@@ -1,13 +1,10 @@
-import type { Event as RBCEvent } from 'react-big-calendar';
-
 export type ReservationStatus = 'confirmed' | 'pending' | 'canceled';
 
-export type CalEvent = RBCEvent & {
+export type CalEvent = {
   id: string;
+  title: string;
+  start: Date;
+  end: Date;
+  status: ReservationStatus;
   place?: string;
-  tone?: 'blue' | 'beige';
-  status?: ReservationStatus;
-  // 필요 시 resource에 원본 객체를 넣을 수 있음 (RBC 기본 필드)
-  // 이게 왜 필요함?
-  resource?: unknown;
 };
